Lisätty kouluttajan arviointien listaus ja arvioinnin antaminen sekä muokkaus
