+++ src/locales/fi.json
@@ -8,6 +8,8 @@
   "arvio-puuttuvasta-koulutuksesta": "Arvio puuttuvasta koulutuksesta",
   "arvioinnin-antaja": "Arvioinnin antaja",
   "arvioinnin-osa-alueita": "Arvioinnin osa-alueita",
+  "arvioinnin-tallentaminen-epaonnistui": "Arvioinnin tallentaminen epäonnistui",
+  "arvioinnin-tallentaminen-onnistui": "Arvioinnin tallentaminen onnistui",
   "arvioinnit": "Arvioinnit",
   "arvioinnit-ja-itsearvioinnit": "Arvioinnit ja itsearvioinnit",
   "arvioinnit-kuvaus": "Voit pyytää arviointia suorituksestasi kouluttajalta, selata arviointeja ja tehdä itsearviointeja.",
@@ -33,6 +35,7 @@
   "arvo": "Arvo",
   "asiakirjat": "Asiakirjat",
   "avaa-arviointi": "Avaa arviointi",
+  "avoin": "Avoin",
   "canmeds-linkki-teksti": "Lue lisää Canmeds osaamisperustaisen lääkärikoulutuksen osa-alueista.",
   "datepicker-label-help": "Käytä nuolinäppäimiä valitsemiseen",
   "datepicker-no-date-selected": "Päivää ei ole valittu",
@@ -55,6 +58,7 @@
   "haluatko-varmasti-poistaa": "Haluatko varmasti poistaa {name}?",
   "helppo": "Helppo",
   "henkilotiedot": "Henkilötiedot",
+  "hyvaksytty": "Hyväksytty",
   "hyvaksytty-toiselle-erikoisalalle": "Hyväksytty toiselle erikoisalalle",
   "hyvin-vaativa": "Hyvin vaativa",
   "itsearvioinnin-tallentaminen-epaonnistui": "Itsearvioinnin tallentaminen epäonnistui",
@@ -135,6 +139,7 @@
   "luottamuksen-taso": "Luottamuksen taso",
   "muokattu": "Muokattu",
   "muokkaa": "Muokkaa",
+  "muokkaa-arviointia": "Muokkaa arviointia",
   "muokkaa-arviointipyyntoa": "Muokkaa arviointipyyntöä",
   "muokkaa-itsearviointia": "Muokkaa itsearviointia",
   "muokkaa-kommenttia": "Muokkaa kommenttia",
@@ -219,6 +224,7 @@
   "tapahtuma": "Tapahtuma",
   "tavallinen": "Tavallinen",
   "tavallista-helpompi": "Tavallista helpompi",
+  "tee-arviointi": "Tee arviointi",
   "tee-itsearviointi": "Tee itsearviointi",
   "tee-tapahtumasta-itsearviointi": "Tee tapahtumasta itsearviointi",
   "teoriakoulutukset": "Teoriakoulutukset",
@@ -226,6 +232,7 @@
   "terveyden-ja-hyvinvoinnin-edistaminen": "Terveyden ja hyvinvoinnin edistäminen",
   "terveyskeskus": "Terveyskeskus",
   "terveyskeskustyo": "Terveyskeskustyö",
+  "tila": "Tila",
   "toimenpiteet": "Toimeenpiteet",
   "toimenpiteita": "Toimenpiteitä",
   "toiminta-epasuoran-ohjauksen-alaisena": "Toiminta epäsuoran ohjauksen alaisena",
@@ -287,6 +294,7 @@
   "valitse-ryhma-enter-nappaimella": "Valitse ryhmä enter-näppäimellä",
   "valittu": "Valittu",
   "valittuna-enimmäismaara": "Valittuna enimmäismäärä: {max}. Poista jokin valittu lisätäksesi uuden.",
+  "valmis": "Valmis",
   "vastuuhenkilo": "Vastuuhenkilö",
   "viestit": "Viestit",
   "virheellinen-syote": "Virheellinen syöte",
