+++ src/main.ts
@@ -8,6 +8,7 @@ import "@/plugins/date";
 import "@/plugins/apexcharts";
 import i18n from "@/plugins/i18n";
 import "@/plugins/fontawesome";
+import "@/plugins/roles";
 import App from "@/app.vue";
 import router from "@/router";
 import store from "@/store";
