+++ src/plugins/fontawesome-regular.ts
@@ -1,12 +1,16 @@
 import { library } from "@fortawesome/fontawesome-svg-core";
 import {
   faCalendarAlt,
+  faCheckCircle,
   faCircle,
   faEnvelope,
-  faHospital
+  faHospital,
+  faClock
 } from "@fortawesome/free-regular-svg-icons";
 
 library.add(faCalendarAlt);
+library.add(faCheckCircle);
 library.add(faCircle);
 library.add(faEnvelope);
 library.add(faHospital);
+library.add(faClock);
